move main layout components into components folder
