+++ src/app/(main)/_components/footer.tsx
