+++ src/app/(main)/layout.tsx
@@ -1,8 +1,7 @@
 import { type ReactNode } from "react";
 import { redirect } from "next/navigation";
- 
-import Header from "./header";
-import { Footer } from "./footer";
+import { Header } from "./_components/header";
+import { Footer } from "./_components/footer";
 import { validateRequest } from "@/lib/auth/validate-request";
 import { redirects } from "@/lib/constants";
